Fund Transfer Functionality
Added Fund Transfer Functionality to the Client Handler. Also,
noticed a static class within the WebServer Class, so I seperated
and made the new class public. Accounts are currently hardcoded,
they need to be read from a file.

Issue: #4
